refactor(bufferThresholds): rename, split modules
